rename salesforcedx-vscode-aura package to -lightning / remove old lightning package
