--- packages/salesforcedx-vscode-aura/.vscodeignore
@@ -1,8 +1,0 @@
-.vscode/**
-.vscode-test/**
-out/test/**
-test/**
-src/**
-**/*.map
-.gitignore
-tsconfig.json
