--- packages/salesforcedx-vscode-aura/src/telemetry/index.ts
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) 2019, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-import { TelemetryService } from './telemetry';
-
-export const telemetryService = TelemetryService.getInstance();
