--- packages/salesforcedx-vscode-aura/src/telemetry/telemetry.ts
@@ -1,59 +1,0 @@
-/*
- * Copyright (c) 2018, salesforce.com, inc.
- * All rights reserved.
- * Licensed under the BSD 3-Clause license.
- * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-
-import * as util from 'util';
-import TelemetryReporter from 'vscode-extension-telemetry';
-
-const EXTENSION_NAME = 'salesforcedx-vscode-lightning';
-
-export class TelemetryService {
-  private static instance: TelemetryService;
-  private reporter: TelemetryReporter | undefined;
-  private isTelemetryEnabled: boolean;
-
-  constructor() {
-    this.isTelemetryEnabled = false;
-  }
-
-  public static getInstance() {
-    if (!TelemetryService.instance) {
-      TelemetryService.instance = new TelemetryService();
-    }
-    return TelemetryService.instance;
-  }
-
-  public initializeService(
-    reporter: TelemetryReporter,
-    isTelemetryEnabled: boolean
-  ): void {
-    this.isTelemetryEnabled = isTelemetryEnabled;
-    this.reporter = reporter;
-  }
-
-  public sendExtensionActivationEvent(hrstart: [number, number]): void {
-    if (this.reporter !== undefined && this.isTelemetryEnabled) {
-      const startupTime = this.getEndHRTime(hrstart);
-      this.reporter.sendTelemetryEvent('activationEvent', {
-        extensionName: EXTENSION_NAME,
-        startupTime
-      });
-    }
-  }
-
-  public sendExtensionDeactivationEvent(): void {
-    if (this.reporter !== undefined && this.isTelemetryEnabled) {
-      this.reporter.sendTelemetryEvent('deactivationEvent', {
-        extensionName: EXTENSION_NAME
-      });
-    }
-  }
-
-  private getEndHRTime(hrstart: [number, number]): string {
-    const hrend = process.hrtime(hrstart);
-    return util.format('%ds %dms', hrend[0], hrend[1] / 1000000);
-  }
-}
