--- packages/salesforcedx-vscode-aura/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs",
-    "target": "es6",
-    "lib": ["dom", "es2016"],
-    "sourceMap": true,
-    "declaration": true,
-    "moduleResolution": "node",
-    "noImplicitAny": true,
-    "rootDir": ".",
-    "outDir": "out",
-    "preserveConstEnums": true,
-    "strict": true,
-    "baseUrl": ".",
-    "paths": {
-      "*": ["node_modules/*", "*"]
-    }
-  },
-  "exclude": ["node_modules", ".vscode-test", "out"]
-}
