+++ packages/salesforcedx-vscode-lightning/LICENSE.txt
@@ -1,4 +1,4 @@
-Copyright (c) 2017, Salesforce.com, inc.
+Copyright (c) 2019, Salesforce.com, inc.
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without modification,
