+++ packages/salesforcedx-vscode-lightning/resources/lwclogo.png
