+++ packages/salesforcedx-vscode-lightning/src/commands/quickpick/quickpick.ts
