+++ packages/salesforcedx-vscode-lightning/src/commands/quickpick/tag-item.ts
