+++ packages/salesforcedx-vscode-lightning/src/views/component-tree-provider.ts
