+++ packages/salesforcedx-vscode-lightning/test/extension.test.ts
