+++ packages/salesforcedx-vscode-lightning/test/index.ts
